Arborist.workspaceNodes and .workspaceDependencySet methods
This adds a simple way for an Arborist instance to turn a list of
workspace names into a list of nodes, or a set of all deps (direct and
transitive) that are relied upon by those workspace nodes.

Note that this returns the *target* of the workspace, not the link in
the root node_modules folder.