Do not add metadata to node if versions mismatch
When a Node is added to a tree, and lacks resolved and integrity values,
but its path exists in the metadata source (shrinkwrap or lockfile), we
apply the metadata that we know to the Node object.  However, we should
_not_ do this when the version doesn't match, because that means it's a
different thing.  In those cases, we should just go ahead and let it
lack the metadata, so that we know to update it later.

This was resulting in very strange cases where the version gets updated,
and the lockfile and hidden lockfile reflect the change, but the actual
package contents in node_modules were not updated, because reify thought
it was the same bits.

Fix: npm/cli#3297
Fix: npm/cli#3196