Properly diff shrinkwrapped nodes without metadata
If a node was included in a shrinkwrap, _but_ that shrinkwrap lacked
resolved and integrity fields, then updating the shrinkwrapping node
would not properly diff the children, treating any updates to them as a
`null` action, even if the version changed.

With this update,

- Any version change is treated as a `CHANGE` action
- Any lacking integrity is treated as a `CHANGE`, unless:
  - neither actual nor ideal have integrity, and
  - resolved is _also_ lacking, and
  - versions match.

That second case (matching versions, no resolved, no integrity) is
important to propery support git dependencies and link targets.

Fix: npm/cli#3196