fix(inventory) handle old and british forms of 'license'
Once upon a time we allowed 'licenses' to provide an array of licenses,
and some packages out in the wild use the British spelling 'licence'.

Also, added a test for the otherwise uncovered behavior of being able to
query for inherited properties, which is important for the packageName
queryable field.