fix: do not try to inflate a fresh lockfile
when running an updateAll, we use Shrinkwrap.reset() to empty and reset
the lockfile metadata. however, due to _inflateAncientLockfile having a
check on meta.originalLockfileVersion, we need that property to be set
on a freshly reset lock so that _inflateAncientLockfile knows it has no
work to be done.

PR-URL: https://github.com/npm/arborist/pull/293
Credit: @nlf
Close: #293
Reviewed-by: @ruyadorno, @isaacs