fix: pass integrity to pacote.extract
when crackOpen becomes true we were extracting the content through
pacote without providing an integrity, because of this pacote was
creating its own internal ssri IntegrityStream that uses the default
algorithm and piping the response data through that for reasons.
when the default algorithm differs from the current one, which is
often the case with older packages that had sha1 digests, we were
throwing an EINTEGRITY error attempting to compare the old and
correct sha1 to a new and unknown sha512 (the default). by passing
the node's current integrity value we ensure that pacote compares
the sha1 to a sha1 which works as expected.

PR-URL: https://github.com/npm/arborist/pull/297
Credit: @nlf
Close: #297
Reviewed-by: @isaacs