feat: make node.target return this for non-links
This gets rid of a lot of `tree = node.target || node` everywhere, by
making `node.target` return `node` when `node` is not a Link.

Caveat is that `if (node.target)` is no longer a thing we can do to test
for link-ness, but it's more clear in those cases to use `if
(node.isLink)` if we want to know if the node is a Link.