feat: add peerEntrySets() function
This returns a Map where the key is an Edge which causes the inclusion
of a set of peers, and the value is the Set of peers.

This enables us to determine whether a set of peers can be replaced,
pushed deeper into the tree, upgraded to new versions, etc., depending
on the source of its inclusion.