CanPlaceDep class
This class takes over the Arborist[_canPlaceDep] function from the
buildIdealTree mixin.

It keeps all state local to the object, which represents a single
attempt to determine whether or not a given Node and its peers can be
placed somewhere within the package tree.  The peerDep checks are also
kept in a tree structure, so that in the case of a conflict, we can have
ready access to all of the problems encountered, which were previously
only being tracked by passing an increasingly long list of arguments
through a chain of function calls.

This class also includes a debug() check to verify that the tree is not
mutated in any way whatsoever in the process of making this
determination.  Note that this debug() check is exceedingly expensive,
but since it only runs in development/test environments, it's a
worthwhile expense.

When integrated, in order to prevent the buildIdealTree test from
becoming hideously slow, debug() mode will be switched off explicitly in
that test.