PlaceDep class
This takes the place of the Arborist[_placeDep] method, and is the
action-taking partner to the CanPlaceDep class.

Given an edge, a node, and its peers, it will attempt to place the Node
and its peers in the tree, making whatever changes it deems necessary.

Like the CanPlaceDep class, this class maintains a tree of child
PlaceDep objects representing the placements of the Node's peers which
were placed along with it.