test: normalize paths and improve tests on Windows
This addresses the issue caused by npm-package-arg in v8.1.5.  While a
correct change, that update had the side effect of calling the
system-specific path.resolve() method on all file: dependency
resolutions, which means that a whole bunch of Arborist's tests have to
be coerced into Unix paths when run on Windows.

Additionally, while debugging this issue, I found that the Shrinkwrap
symbolic link tests do not work on Windows unless they are run in a mode
where normal (non-Junction) symlinks are allowed, so that is addressed
as well.