fix(reify): verify existing deps in nm are dirs
Verify that an existing folder that might exist in the
`createSparseTree` phase is actually a directory and not merely a
symbolic link to one (or something else).  Note that this can only occur
on case-sensitive filesystems where the folder names differ only by
case, or else it would have already been moved aside.

This also walks up the folder tree ensure that we're not in a case where
the sparse leaf _parent_ is a symlink.