fix: node_modules must be a directory
This adds a check to the extraction process to ensure that the
node_modules into which a node is being extracted is a true directory,
and not a symbolic link.

This is performed at every level, so that the symlink cannot be hidden
within a workspace, or created via some other creative means.

If `--force` is set, then the safety measure is disabled.