fix: make Node.children() a case-insensitive Map
In other words, do not allow a node to have both `FOO` and `foo` in a
parent node's `children` collection.

In order to maintain proper semantics for package resolution, this also
means that Node.edgesOut has to be case-insensitive as well, meaning
that a package may not *depend* on both `FOO` and `foo`.

Lastly, do not add removed dependency folders to the trashlist.  There
is no need, since they are retired and then not un-retired, and removing
the package folder itself (which was always a rimraf of a non-existent
path) could remove something we want, if you are replacing `FOO` with
`foo`.

This has the effect of making Arborist.buildIdealTree() and
Arborist.reify() properly collide packages that differ only in case,
thus generating trees that are safe for use across a variety of file
systems.

This is an improvement on the previous commit's "non-directory detected
during sparse tree creation" check, because it will avoid cases where
the package-lock.json does not match what we actually end up creating on
disk, and sidestep the collisions altogether.

Note that it _is_ still possible to end up with collisions on Windows
systems, if for example you have a package.json like this:

```json
{
  "dependencies": {
    "longer-than-eight-char-name": "file:/some/arbitrary/path",
    "LONGER~1": "npm:arbitrary-write@*"
  }
}
```

In that case, the sparse dir check in the previous commit will be
triggered, since the folder's non-directory-ness will be detected while
creating the sparse folder tree.  The user will end up with a mismatched
lockfile, but importantly, they *won't* have arbitrary writes being sent
through a symbolic link.

On case _sensitive_ systems, this means that the `node_modules` folder
may not contain case-variant packages, and any projects currently in
this state will need to be reinstalled or updated.  As this is extremely
rare (since uppercase and non-url-safe characters have not been allowed
in package names on the public registry for many years), it is unlikely
to affect any users in the wild.