reify: debug crash when extracting into symlink
We should never be doing this.  Any time we're extracting contents, it
should be into an honest to goodness directory, and not any other kind
of thing.

Commits to follow will make it impossible to reach this condition
without some very deep hacking, but the check is here to make it easier
to debug if anything like this is reported again.

Related-to: https://github.com/npm/cli/issues/3457