fix: load global symlinks implicitly as file: deps
When running `loadActual` or `buildIdealTree()` in the global root, we
read all or some of the top-level installed packages, and create a
"pseudo-package" for the root node which depends on all of them with the
range '*'.

However, for *linked* packages, this is not quite accurate.  In that
case, we really do not want to change them by accident to anything other
than the actual symlink we put there.

Detect this state, and assign the dependency properly in the root level.

Note that this does _not_ fix https://github.com/npm/cli/issues/3457,
but it did contribute to attempting to ending up with a Node rather than
a Link in the top-level of the global install space.  With this patch in
place, the condition described will result in an `ERESOLVE` error rather
than destroying the link target, but that is still arguably preferrable.

Related-to: https://github.com/npm/cli/issues/3457