fix(test): Some fixes for skipped tests and path normalization
On case-sensitive filesystems, we were skipping a test, but in so doing,
not replacing the clobbered Arborist method, which made subsequent tests
fail.

Also, not normalizing the paths in the set of warnings, resulting in a
spurious failure on Windows.

Finally, a duplicated test is removed, which was added by mistake while
resolving a merge conflict.