fix: set lockfileVersion from file during reset
Fix: npm/cli#3920

When loading the initial tree while updating all, a shrinkwrap file
loaded from disk with an existing `lockfileVersion` was not having the
`lockfileVersion` preserved when saving the file back to disk.

With this patch, an existing `lockfileVersion` is preserved if it is
greater than the `defaultLockfileVersion`.

Co-authored-by: @isaacs

PR-URL: https://github.com/npm/arborist/pull/340
Credit: @lukekarrys
Close: #340
Reviewed-by: @isaacs