fix: stop pruning peerSets when entryEdge is from a workspace
Fix: npm/cli#3933

Previously, installing a tree that causes peerSets to be pruned would
cause an infinite loop when done inside a workspace. This patch stops
that check if the entryEdge is a workspace, the same as being done if it
were the project root.

This patch also adds some incorrect tests that were discovered when
debugging this, and will be fixed in the future (#347 #348).

PR-URL: https://github.com/npm/arborist/pull/343
Credit: @lukekarrys
Close: #343
Reviewed-by: @isaacs