fix: normalize path used as cache key
Normalize path to avoid mixed path separators on
Windows causing a cache miss for paths that refer
to the same thing. This would otherwise cause a
new node with the path of an existing node to be
created which in turns deletes the existing node
and its children. However, the children will not
be loaded again from the file system due to the
_actualTreeLoaded check in _loadFSTree.